fix: hide DO_NOT_BUNDLE
