+++ e2e/fixtures/rsc-basic/waku.config.ts
@@ -1,15 +1,7 @@
-const DO_NOT_BUNDLE = '';
-
 /** @type {import('waku/config').Config} */
 export default {
-  middleware: (cmd: 'dev' | 'start') => [
-    ...(cmd === 'dev'
-      ? [
-          import(
-            /* @vite-ignore */ DO_NOT_BUNDLE + 'waku/middleware/dev-server'
-          ),
-        ]
-      : []),
+  middleware: () => [
+    import('waku/middleware/dev-server'),
     import('waku/middleware/rsc'),
     import('waku/middleware/fallback'),
   ],
