+++ packages/waku/src/lib/config.ts
@@ -16,12 +16,8 @@ const ADDITIONAL_HTML_HEAD = `
 <meta name="generator" content="Waku" />
 `.trim();
 
-const DO_NOT_BUNDLE = '';
-
-const DEFAULT_MIDDLEWARE = (cmd: 'dev' | 'start') => [
-  ...(cmd === 'dev'
-    ? [import(/* @vite-ignore */ DO_NOT_BUNDLE + 'waku/middleware/dev-server')]
-    : []),
+const DEFAULT_MIDDLEWARE = () => [
+  import('waku/middleware/dev-server'),
   import('waku/middleware/headers'),
   import('waku/middleware/ssr'),
   import('waku/middleware/rsc'),
